Добавил проверки на авторизацию для запросов через токен

+++ routes/record.routes.js
@@ -1,13 +1,24 @@
 const {Router} = require("express");
 const Record = require("../models/Record");
+const {verifyToken} = require("../utils/tokenUtils");
+const {NotAuthorizedError} = require("../utils/errorUtils");
+const jwt = require("jsonwebtoken");
+const config = require("config");
 
 const router = new Router();
 
 // /api/record/create
 router.post(
     "/create",
+    verifyToken,
     async (req, res) => {
         try {
+            const decoded = await jwt.verify(req.token, config.get("jwtSecret"));
+
+            if (decoded.userId !== req.params.id) {
+                throw new NotAuthorizedError();
+            }
+
             const record = new Record(req.body);
             const result = await record.save();
 
@@ -15,7 +26,13 @@ router.post(
         } catch (error) {
             console.log('Error:', error.message);
 
-            res.status(500).json({ message: "Что-то пошло не так, попробуйте снова" })
+            if (error.name === "TokenExpiredError") {
+                res.status(403).json({message: "Сессия истекла."});
+            } else if (error.name === "NotAuthorizedError") {
+                res.status(403).json({message: error.message});
+            } else {
+                res.status(500).json({message: "Что-то пошло не так, попробуйте снова"});
+            }
         }
     }
 );
@@ -23,15 +40,28 @@ router.post(
 // /api/record/:id/update
 router.put(
     "/:id/update",
+    verifyToken,
     async (req, res) => {
         try {
+            const decoded = await jwt.verify(req.token, config.get("jwtSecret"));
+
+            if (decoded.userId !== req.params.id) {
+                throw new NotAuthorizedError();
+            }
+
             const record = await Record.findByIdAndUpdate(req.params.id, {$set: req.body}, {useFindAndModify: false, new: true}).exec();
 
             res.status(201).json(record);
         } catch (error) {
             console.log('Error:', error.message);
 
-            res.status(500).json({message: "Что-то пошло не так, попробуйте снова"})
+            if (error.name === "TokenExpiredError") {
+                res.status(403).json({message: "Сессия истекла."});
+            } else if (error.name === "NotAuthorizedError") {
+                res.status(403).json({message: error.message});
+            } else {
+                res.status(500).json({message: "Что-то пошло не так, попробуйте снова"});
+            }
         }
     }
 );
@@ -39,15 +69,28 @@ router.put(
 // /api/record/:id/delete
 router.delete(
     "/:id/delete",
+    verifyToken,
     async (req, res) => {
         try {
+            const decoded = await jwt.verify(req.token, config.get("jwtSecret"));
+
+            if (decoded.userId !== req.params.id) {
+                throw new NotAuthorizedError();
+            }
+
             await Record.findOneAndDelete({'_id' : req.params.id}).exec();
 
             res.status(201).json({message: "Запись успешно удалена"});
         } catch (error) {
             console.log('Error:', error.message);
 
-            res.status(500).json({message: "Что-то пошло не так, попробуйте снова"})
+            if (error.name === "TokenExpiredError") {
+                res.status(403).json({message: "Сессия истекла."});
+            } else if (error.name === "NotAuthorizedError") {
+                res.status(403).json({message: error.message});
+            } else {
+                res.status(500).json({message: "Что-то пошло не так, попробуйте снова"});
+            }
         }
     }
 );
